Initial import of smstools3 3.1.20

+++ src/locking.c
@@ -3,8 +3,7 @@ SMS Server Tools 3
 Copyright (C) 2006- Keijo Kasvi
 http://smstools3.kekekasvi.com/
 
-Based on SMS Server Tools 2 from Stefan Frings
-http://www.meinemullemaus.de/
+Based on SMS Server Tools 2, http://stefanfrings.de/smstools/
 SMS Server Tools version 2 and below are Copyright (C) Stefan Frings.
 
 This program is free software unless you got it under another license directly
@@ -45,14 +44,19 @@ int lockfile( char*  filename)
       // 3.1.15:
       //snprintf(pid, sizeof(pid), "%i %s\n", (int)getpid(), DEVICE.name);
       snprintf(pid, sizeof(pid), "%i %s\n", (int)getpid(),
-               (process_id == -1) ? "MAINPROCESS" : DEVICE.name);
+               // 3.1.18: Use process_title now when there are other process_id's below 0,
+               // even when CHILD or NOTIFIER do not use lockfiles (at least currently):
+               //(process_id == -1) ? "MAINPROCESS" : DEVICE.name);
+               process_title);
 
       write(lockfile, pid, strlen(pid));
+      // 3.1.16beta: Fix: Use fsync instead of sync after close:
+      fsync(lockfile);
       close(lockfile);
-      sync();
       return 1;
     }
   }
+
   return 0;
 }
 
@@ -76,6 +80,7 @@ int islocked( char*  filename)
 int unlockfile( char*  filename)
 {
   char lockfilename[PATH_MAX +5];
+  struct stat statbuf;
 
   if (!filename)
     return 0;
@@ -84,7 +89,8 @@ int unlockfile( char*  filename)
 
   strcpy(lockfilename,filename);
   strcat(lockfilename,".LOCK");
-  if (unlink(lockfilename))
-    return 0;
+  if (!stat(lockfilename, &statbuf)) // 3.1.16beta: Check if file exists
+    if (unlink(lockfilename))
+      return 0;
   return 1;
 }
